requirements.txt
wxPython
pdfminer.six
pdf2docx
pymupdf
